Floating Panel incomplete addition
The Floating Panel is an item that works like an glass hole: it doesn't get affected if the ground goes up or down. It can be used to make the chamber look more futuristic, and I guess that's it. It's a floating platform. Got this idea from a Portal fangame anyways lol